refactor(naming): SCREAMING_SNAKE 변환 헬퍼 추가 및 ORM 백엔드 통합
